requirements.txt
accelerate==0.30.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.34.1
antlr4-python3-runtime==4.9.3
anyio==4.4.0
attrs==24.2.0
bitsandbytes==0.44.1
blis==0.7.11
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools==5.5.0
catalogue==2.0.10
certifi @ file:///croot/certifi_1720453481653/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.8
datasets==2.21.0
deepspeed==0.14.4
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.0
eval_type_backport==0.2.0
fastapi==0.114.0
fastchat==0.1.0
filelock @ file:///work/perseverance-python-buildout/croot/filelock_1701733993137/work
fire==0.7.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gguf==0.9.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
groq==0.11.0
grpcio==1.66.1
grpcio-status==1.62.3
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.24.6
humanfriendly==10.0
hydra-core==1.3.2
idna @ file:///croot/idna_1714398848350/work
importlib_metadata==8.4.0
interegular==0.3.3
jaxtyping==0.2.34
Jinja2 @ file:///croot/jinja2_1716993405101/work
jiter==0.5.0
joblib==1.4.2
jsonlines==4.0.0
jsonpath-python==1.0.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
lark==1.2.2
litellm==1.49.6
llvmlite==0.43.0
lm-format-enforcer==0.10.6
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe @ file:///work/perseverance-python-buildout/croot/markupsafe_1707342994941/work
matplotlib==3.9.2
mdurl==0.1.2
mistral_common==1.3.4
mistralai==1.1.0
mkl-fft @ file:///work/perseverance-python-buildout/croot/mkl_fft_1698845673361/work
mkl-random @ file:///work/perseverance-python-buildout/croot/mkl_random_1698845720894/work
mkl-service==2.4.0
mpmath @ file:///work/perseverance-python-buildout/croot/mpmath_1698864994882/work
msgpack==1.0.8
msgspec==0.18.6
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx @ file:///croot/networkx_1720002482208/work
ninja==1.11.1.1
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.161
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
omegaconf==2.3.0
onnxruntime==1.19.2
openai==1.51.2
outlines==0.0.46
packaging==24.1
pandas==2.2.2
partial-json-parser==0.2.1.1.post4
peft==0.12.0
pillow @ file:///croot/pillow_1721059439630/work
platformdirs==4.3.6
preshed==3.0.9
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycountry==24.6.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.4
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML @ file:///work/perseverance-python-buildout/croot/pyyaml_1698849903511/work
pyzmq==26.2.0
ray==2.35.0
referencing==0.35.1
regex==2024.7.24
requests @ file:///croot/requests_1721410876868/work
rich==13.8.0
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.1
sentencepiece==0.2.0
sentry-sdk==2.14.0
setproctitle==1.3.3
setuptools==72.1.0
shellingham==1.5.4
shtab==1.7.1
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.38.5
sympy @ file:///croot/sympy_1724938189289/work
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.4.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
together==1.2.10
tokenizers==0.19.1
torch==2.4.0
torchaudio==2.3.1
torchvision==0.19.0
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0
trl==0.8.6
typeguard==2.13.3
typer==0.9.4
typing-inspect==0.9.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tyro==0.8.11
tzdata==2024.1
uritemplate==4.1.1
urllib3 @ file:///croot/urllib3_1718912636303/work
uvicorn==0.30.6
uvloop==0.20.0
vllm==0.6.0
vllm-flash-attn==2.6.1
wandb==0.18.1
wasabi==1.1.3
watchfiles==0.24.0
weasel==0.3.4
websockets==13.0.1
wheel==0.43.0
wrapt==1.16.0
xformers==0.0.27.post2
xgboost==2.1.1
xxhash==3.5.0
yarl==1.9.7
zhipuai==2.1.5.20230904
zipp==3.20.1
